weather dynamic fixes

+++ src/component/Map.js
@@ -520,10 +520,12 @@ class Favoris extends Component {
   constructor(props) {
     super(props);
     this.state = {
-    favorites: []
+    favorites: [],
+    favoritesWeatherDatas: []
     };
     this.closeComponent = this.closeComponent.bind(this);
     this.deleteFavorite = this.deleteFavorite.bind(this);
+
   }
 
   closeComponent(){
@@ -541,33 +543,40 @@ class Favoris extends Component {
     return response.json();
     }).then(function(data) {
       var userFavorites = data.favorites
-      console.log('data favorites', data.favorites);
       ctx.setState({
         favorites:userFavorites
       })
-      for (var i = 0; i < data.favorites.length; i++) {
-        console.log(data.favorites[i]);
+
+      userFavorites.map((favorite, i) => {
+        fetch('http://localhost:3000/getLocationWeatherInfos', {
+        method: 'POST',
+        headers: {'Content-Type':'application/x-www-form-urlencoded'},
+        body: 'latitude='+favorite.latitude+'&longitude='+favorite.longitude
+        })
+        .then(function(response) {
+        return response.json();
+        })
+        .then(function(weatherData) {
+          console.log('weatherData', weatherData);
+
+        var favorites = [...ctx.state.favorites];
+        console.log('favorites', favorites);
+        favorites[i].weatherDatas = weatherData;
+        ctx.setState({
+          favorites
+        })
+
+        });
+      })
 
 
-    //     fetch('http://localhost:3000/getLocationWeatherInfos', {
-    //     method: 'POST',
-    //     headers: {'Content-Type':'application/x-www-form-urlencoded'},
-    //     body: 'latitude='+data[i].latitude+'&longitude='+data[i].longitude
-    //     })
-    // .then(function(response) {
-    //     return response.json();
-    // })
-    // .then(function(weatherData) {
-    //     console.log('weather data favorite: ', weatherData);
-    //     var weatherDataCopy = {...weatherData}
-    // });
-
-      }
 
     });
     }
 
-    componentDidUpdate(){
+    componentDidUpdate(prevProps){
+      if (this.props.userId !== prevProps.userId) {
+        console.log('tata');
       const ctx= this;
       fetch('http://localhost:3000/favorites', {
       method: 'POST',
@@ -582,6 +591,7 @@ class Favoris extends Component {
         favorites:userFavorites
       })
       });
+      }
     }
 
 deleteFavorite(locationName) {
@@ -592,13 +602,17 @@ deleteFavorite(locationName) {
   body: 'userId='+this.props.userId+'&locationName='+locationName
   })
   .then(function(response) {
-    console.log('Response delete', response );
   return response.json();
   }).then(function(data) {
-    console.log('data delete: ', data );
-    var userFavorites = data.user.favorite;
+    var userFavorites = data.deleteId;
+    var favoriteCopy = [...ctx.state.favorites];
+    for (var i = 0; i < favoriteCopy.length; i++) {
+      if (favoriteCopy[i].locationName == data.name) {
+        favoriteCopy.splice(i, 1);
+      }
+    }
     ctx.setState({
-      favorites:userFavorites
+      favorites:favoriteCopy
     })
 
   });
@@ -608,19 +622,24 @@ deleteFavorite(locationName) {
 
 
   render() {
-var ctx = this;
-var favoritesList = ctx.state.favorites.map(
-  function(data){
-    return(
-      <Col className="favItem" xs="11" sm="8" md={{ size: 8 }}>{data.locationName}
-        <FontAwesomeIcon className="iconStyle" icon={faSun}/>
-        <h6 className="favFont">Météo actuelle</h6>
-        <p>Ciel dégagé, 25°C, Brise légère, 2.6 m/s</p>
-        <FontAwesomeIcon onClick={()=>ctx.deleteFavorite( data.locationName )} className="iconStyle" icon={faTimesCircle} />
-        </Col>
-    )
+
+  var ctx = this;
+  var favoritesList = ctx.state.favorites.map(
+    function(data, i){
+      console.log('map data',data);
+      if (data.weatherDatas  &&  data.weatherDatas.weather && data.weatherDatas.weather[0]) {
+      return(
+        <Col className="favItem" xs="11" sm="8" md={{ size: 8 }}>{data.locationName}
+          <img src={"http://openweathermap.org/img/w/" + data.weatherDatas.weather[0].icon + ".png"}/>
+          <h6 className="favFont">Météo actuelle</h6>
+          <p>{data.weatherDatas.weather[0].description}, {data.weatherDatas.main.temp} °C</p>
+          <FontAwesomeIcon onClick={()=>ctx.deleteFavorite( data.locationName )} className="iconStyle" icon={faTimesCircle} />
+          </Col>
+      )
+    }
   }
-  )
+    )
+
     return (
 
       <div className="background">
